+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './useTable';
 export * from './types';
 export * from './plugins';
+export * from './strategies';